# typescript-templates/typescript-webapi/src/models/helloWorld.ts
export class HelloWorld {
    /**
     * @example
     * {
     *  "helloWorld": "Hello World",
     *  "id": 1,
     *  "isIndexed": true
     * }
     */
    public helloWorld: string;
    public id: number;
    public isIndexed?: boolean;
}